Further fixes
I had tried to use mutex and .tryLock() to prevent "Start Measurement" button being used during measurements.
It's possible to disable this button, so the measurement thread disables the "Start Measurement" button and renables it as needed.
Much safter than messing around with .tryLock(), which as I found out, doesn't just check if the mutex is locked but will also lock it in the event of it being free.